[PBCKP-304] auto tests cleanup added to ptrack_helper.py
